+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## [Unreleased]
 ### Changed
 - Re-implemented with zustand as a dependency
-- BREAKING CHANGE: `createStore` returns `useStoreState`
-  - `useGlobalState` returned by `createStore` is deprecated.
+- `createStore` returns `useStoreState`
+  - BREAKING CHANGE: `useGlobalState` returned by `createStore` is deprecated
 ### Removed
 - BREAKING CHANGE: drop reduxDevToolsExt
   - Reference implementation in `examples/08_thunk/src/devtools.ts`